==> response/notebooks/stage0_unpack_data.ipynb <==
# unpack_data

This notebook:
1. Solve the password puzzle
2. Use the password to unpack two zip files: `test_data.db.zip` and `test_data.zip` into `response/data/` folder

# Imports

import hashlib

# Copy the ZIP files over to response/data/

!ls ../../data/sample-data/

!cp ../../data/sample-data/test_data* ../data/compressed

!ls ../data/compressed/

# Solve the password puzzle

raw_password = b"welcometotheiconic"

decoded_password = hashlib.sha256(raw_password).hexdigest()

decoded_password

# Unzip the files

!unzip -P $decoded_password -d ../data/raw ../data/compressed/test_data.zip

!unzip -P $decoded_password -d ../data/raw ../data/compressed/test_data.db.zip

!ls -l ../data/raw

# All done!
