==> response/notebooks/stage2_clean.ipynb <==
# stage2_clean
Instructions:
> Unhash the data (/data/sample-data/test_data.zip) using the secret key provided by us, extract it, most importantly clean it and put it in a form you can use - all programatically of course. We have also "intentionally" corrupted two columns in this file - two columns that might look correct but are not correct. They need "some correction" to be useful.

# Imports

import pandas as pd
import re
import io
import traceback

# Ingestion

raw_json_file = "../data/raw/data.json"

# went back to add try except so that the error doesn't interrupt Run All for this notebook
try:
    pd.read_json(raw_json_file)
except Exception as e:
    print(traceback.format_exc())

Come on.. do we really need to corrupt the json file too?

with open(raw_json_file) as f:
    raw_json_str = ''.join(f.readlines())

print(raw_json_str[:1000])

Looks like multiple root level items and no comma separation

separator_pattern = r'\}\n\{'
re.findall(separator_pattern, raw_json_str)[:10]

fixed_json_str = "[" + re.sub(separator_pattern, '},\n{', raw_json_str) + "]"

print(fixed_json_str[:1000])

with io.StringIO(fixed_json_str) as f:
    input_data = pd.read_json(f)

input_data

input_data.info()

Now that's fixed, let me inspect the data

# Inspect

input_data.describe()

This is not helping at all, let me borrow the column ordering from the sqlite db

columns_ordered = ['customer_id', 'days_since_first_order', 'days_since_last_order',
       'is_newsletter_subscriber', 'orders', 'items', 'cancels', 'returns',
       'different_addresses', 'shipping_addresses', 'devices', 'vouchers',
       'cc_payments', 'paypal_payments', 'afterpay_payments', 'apple_payments',
       'female_items', 'male_items', 'unisex_items', 'wapp_items',
       'wftw_items', 'mapp_items', 'wacc_items', 'macc_items', 'mftw_items',
       'wspt_items', 'mspt_items', 'curvy_items', 'sacc_items', 'msite_orders',
       'desktop_orders', 'android_orders', 'ios_orders', 'other_device_orders',
       'work_orders', 'home_orders', 'parcelpoint_orders',
       'other_collection_orders', 'redpen_discount_used',
       'coupon_discount_applied', 'average_discount_onoffer',
       'average_discount_used', 'revenue']

with pd.option_context('display.max_columns', 100):
    display(input_data[columns_ordered].describe())

I already discovered __days_since_last_order__ was actually "hours" in Stage 1 challenge. 

The other "error" was quite obvious as well, __average_discount_used__ should be on the same scale as average_discount_onoffer. Conveniently the max for average_discount_used is exactly 10000, so it's reasonable to simply assume it's out by 4 decimal points. 

Now let's clean!

input_data_clean = (
    input_data[columns_ordered]
    .assign(days_since_last_order=lambda df: df.days_since_last_order/24.0)
    .assign(average_discount_used=lambda df: df.average_discount_used/10000.0)
)

with pd.option_context('display.max_columns', 100):
    display(input_data_clean.describe())

Yep, that looks right. 

Saving the data for the next stage

output_file_path = '../data/processed/clean_data.parquet'
input_data_clean.to_parquet(output_file_path)
