==> response/notebooks/stage3_build_feature_engineering.ipynb <==
# stage_3_build_feature_engineering
This notebook transforms the cleaned data from stage 2 and generates a set of all numeric (float) features, then save it to the `data/processed` folder. 

Conscious of time, I'm not going to experiment with complex feature engineering ideas, rather just focus on 2 things:
- Convert life time transactional metrics like: orders to date, items to date etc., into orders/items per month
- Convert categorical features into numeric if they aren't numeric already

__After I figure out the Feature Engineering steps in this notebook, I'll refactor the pipeline into a .py file under `/response/src` and write a simple unit test. So that the feature engineering pipeline can be imported and called upon by all following stages.__

# Imports

import pandas as pd
import numpy as np

# Load cleaned data from stage2

input_data_path = '../data/processed/clean_data.parquet'
input_data = pd.read_parquet(input_data_path)

with pd.option_context('display.max_columns', 999):
    display(input_data.sample(20))

# Calculate customer tenure

input_data = input_data.assign(
    tenure_months=lambda df: (df.days_since_first_order - df.days_since_last_order) / 30.0
)

In a naive way, I simply calculated the gap between first and last order and divided by 30 days, hope it's a good enough proxy for tenure

input_data.tenure_months.describe()

Small tenure of 0 or 0.03 wouldn't work, in division it will result in huge numbers or inifinity

I'll just round them up to the closest integer

input_data['tenure_months'] = np.ceil(input_data.tenure_months)

input_data['tenure_months'] = input_data['tenure_months'].replace(0, 1.0)

input_data.tenure_months.describe()

That looks much better

# Convert transactional features into monthly

features_numeric_trasactional = input_data[['orders','items','cancels','returns',
                                            'vouchers',
                                        'female_items','male_items','unisex_items','wapp_items','wftw_items',
                                        'mapp_items','wacc_items','macc_items','mftw_items','wspt_items','mspt_items',
                                        'curvy_items','sacc_items',
                                        'msite_orders','desktop_orders','android_orders','ios_orders','other_device_orders',
                                        'work_orders','home_orders','parcelpoint_orders','other_collection_orders',
                                        'redpen_discount_used','coupon_discount_applied',
                                        'revenue',
                                       ]].fillna(0.0)


Divide those feature by tenure_months

features_numeric_monthly = (1.0 * features_numeric_trasactional).div(input_data.tenure_months, axis='index')

features_numeric_monthly.describe()

features_numeric_monthly.info()

Looks good

# Get other numeric features

features_numeric_other = input_data[['days_since_first_order','days_since_last_order','tenure_months',
                                     'different_addresses','shipping_addresses','devices',
                                 'average_discount_onoffer','average_discount_used',
                                ]]


features_numeric_other.info()

Convert everything to float:

features_numeric_other = features_numeric_other * 1.0

features_numeric_other.info()

Done

# Handle the categorical features

features_categorical = input_data[['is_newsletter_subscriber', 'cc_payments','paypal_payments','afterpay_payments','apple_payments']]

features_categorical

Convert Y/N to float

features_categorical = features_categorical\
.assign(is_newsletter_subscriber=lambda df: (df.is_newsletter_subscriber == 'Y') * 1.0)

features_categorical

Everything else to float as well:

features_categorical = 1.0 * features_categorical

features_categorical.info()

# Conbine the feature sets

Put customer_id in there in case we need addtional index or reference

features = pd.concat([
    input_data[['customer_id']], 
    features_categorical, 
    features_numeric_monthly, 
    features_numeric_other
], 
    axis = 1
)

features

features.isnull().any()[lambda s: s]

Coupon_discount_applied still has null values in it, must have missed it somehow..

features = features.fillna(0.0)

features.isnull().any()[lambda s: s]

All good now

Now just double checking there's no infinity in the dataframe:

features.iloc[:, 1:].max().max()

All good to go

# Refactor feature engineering into Python file

I want the Feature Engineering pipeline to live in a .py file instead of this notebook, so that it can be shared across the later stages by importing. 

To achieve this I'll need some data to create a simple unit test. 

Concretely I'll sample one hundred records from `clean_data.parquet` (input) and correspondong one hundred records from `features` dataframe (outout). The unit test for my feature engineering pipeline would be: given the input, it should produce the exact same output. Therefore I know it's functioning identical to this notebook. 

The refactored code will live in `/response/src/feature_engineering.py`, the unit test and test data will be saved in `/response/test/`

test_input = input_data.sample(100, random_state=273)

test_input

test_output = features.loc[test_input.index]

test_output

test_input.to_parquet("../test/fe_test_input.parquet")
test_output.to_parquet("../test/fe_test_output.parquet")
